requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
requests>=2.31.0
arxiv>=2.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
pydantic>=2.6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
groq>=0.4.0
ddgs